# static/swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Incidenta API documentation.",
    "title": "Incidenta API",
    "version": "0.1.0"
  },
  "host": "127.0.0.1:8080",
  "paths": {
    "/v1/alert": {
      "get": {
        "description": "List operation",
        "tags": [
          "Alert"
        ],
        "operationId": "ListAlert",
        "responses": {
          "200": {
            "description": "Alert",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Alert"
              }
            }
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/alert/{alert_id}": {
      "get": {
        "description": "Get operation",
        "tags": [
          "Alert"
        ],
        "operationId": "GetAlert",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "AlertID",
            "name": "alert_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Alert",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      },
      "delete": {
        "description": "Delete operation",
        "tags": [
          "Alert"
        ],
        "operationId": "DeleteAlert",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "AlertID",
            "name": "alert_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/alert/{alert_id}/logs": {
      "get": {
        "description": "Get logs operation",
        "tags": [
          "Alert"
        ],
        "operationId": "ListAlertLogs",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "AlertID",
            "name": "alert_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Log",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Log"
              }
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/alertmanager/webhook": {
      "post": {
        "description": "Alertmanager Webhook operation",
        "tags": [
          "Alertmanager"
        ],
        "operationId": "AlertmanagerWebhook",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "201": {},
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/log/{log_id}": {
      "get": {
        "description": "Get operation",
        "tags": [
          "Log"
        ],
        "operationId": "GetLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "LogID",
            "name": "log_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Alert",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      },
      "delete": {
        "description": "Delete operation",
        "tags": [
          "Log"
        ],
        "operationId": "DeleteLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "LogID",
            "name": "log_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/receiver": {
      "get": {
        "description": "List operation",
        "tags": [
          "Receiver"
        ],
        "operationId": "ListReceiver",
        "responses": {
          "200": {
            "description": "Receiver",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Receiver"
              }
            }
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      },
      "post": {
        "description": "Create operation",
        "tags": [
          "Receiver"
        ],
        "operationId": "CreateReceiver",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ReceiverCreateOptions"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Receiver",
            "schema": {
              "$ref": "#/definitions/Receiver"
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    },
    "/v1/receiver/{receiver_id}": {
      "get": {
        "description": "Get operation",
        "tags": [
          "Receiver"
        ],
        "operationId": "GetReceiver",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ReceiverID",
            "name": "receiver_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Receiver",
            "schema": {
              "$ref": "#/definitions/Receiver"
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      },
      "post": {
        "description": "Edit operation",
        "tags": [
          "Receiver"
        ],
        "operationId": "EditReceiver",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ReceiverID",
            "name": "receiver_id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ReceiverEditOptions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receiver",
            "schema": {
              "$ref": "#/definitions/Receiver"
            }
          },
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      },
      "delete": {
        "description": "Delete operation",
        "tags": [
          "Receiver"
        ],
        "operationId": "DeleteReceiver",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ReceiverID",
            "name": "receiver_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/responses/GenericError"
          },
          "404": {
            "$ref": "#/responses/GenericError"
          },
          "500": {
            "$ref": "#/responses/GenericError"
          }
        }
      }
    }
  },
  "definitions": {
    "Alert": {
      "type": "object",
      "properties": {
        "body": {
          "type": "string",
          "x-go-name": "Body"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedUnix"
        },
        "hash": {
          "type": "string",
          "x-go-name": "Hash"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "receiver_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReceiverID"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedUnix"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "Alerts": {
      "type": "array",
      "title": "Alerts is a list of Alert objects.",
      "items": {
        "$ref": "#/definitions/Alert"
      },
      "x-go-package": "github.com/prometheus/alertmanager/template"
    },
    "KV": {
      "type": "object",
      "title": "KV is a set of key/value string pairs.",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/prometheus/alertmanager/template"
    },
    "Log": {
      "type": "object",
      "properties": {
        "alert_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AlertID"
        },
        "comment": {
          "type": "string",
          "x-go-name": "Comment"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedUnix"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "receiver_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReceiverID"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedUnix"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "Message": {
      "type": "object",
      "title": "Message defines the JSON object send to webhook endpoints.",
      "properties": {
        "alerts": {
          "$ref": "#/definitions/Alerts"
        },
        "commonAnnotations": {
          "$ref": "#/definitions/KV"
        },
        "commonLabels": {
          "$ref": "#/definitions/KV"
        },
        "externalURL": {
          "type": "string",
          "x-go-name": "ExternalURL"
        },
        "groupKey": {
          "type": "string",
          "x-go-name": "GroupKey"
        },
        "groupLabels": {
          "$ref": "#/definitions/KV"
        },
        "receiver": {
          "type": "string",
          "x-go-name": "Receiver"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "version": {
          "description": "The protocol version.",
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/prometheus/alertmanager/notify/webhook"
    },
    "Receiver": {
      "type": "object",
      "properties": {
        "ack_button": {
          "type": "boolean",
          "x-go-name": "AckButton"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedUnix"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "resolve_button": {
          "type": "boolean",
          "x-go-name": "ResolveButton"
        },
        "slack_url": {
          "type": "string",
          "x-go-name": "SlackURL"
        },
        "snooze_button": {
          "type": "boolean",
          "x-go-name": "SnoozeButton"
        },
        "template_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TemplateID"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedUnix"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "ReceiverCreateOptions": {
      "type": "object",
      "required": [
        "name",
        "slack_url",
        "template_id"
      ],
      "properties": {
        "ack_button": {
          "type": "boolean",
          "x-go-name": "AckButton"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "resolve_button": {
          "type": "boolean",
          "x-go-name": "ResolveButton"
        },
        "slack_url": {
          "type": "string",
          "x-go-name": "SlackURL"
        },
        "snooze_button": {
          "type": "boolean",
          "x-go-name": "SnoozeButton"
        },
        "template_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TemplateID"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "ReceiverEditOptions": {
      "type": "object",
      "properties": {
        "ack_button": {
          "type": "boolean",
          "x-go-name": "AckButton"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "resolve_button": {
          "type": "boolean",
          "x-go-name": "ResolveButton"
        },
        "slack_url": {
          "type": "string",
          "x-go-name": "SlackURL"
        },
        "snooze_button": {
          "type": "boolean",
          "x-go-name": "SnoozeButton"
        },
        "template_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TemplateID"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "Template": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedUnix"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedUnix"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "TemplateCreateOptions": {
      "type": "object",
      "required": [
        "name",
        "content"
      ],
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    },
    "TemplateEditOptions": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/incidenta/incidenta/pkg/api"
    }
  },
  "responses": {
    "GenericError": {
      "headers": {
        "error": {
          "type": "string"
        },
        "error_description": {
          "type": "string"
        }
      }
    }
  }
}